refactor into adapters.camera package
