+++ app/src/main/java/org/WenuLink/adapters/aircraft/AircraftHandler.kt
@@ -13,8 +13,8 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
 import org.WenuLink.adapters.AsyncUtils
-import org.WenuLink.adapters.CameraHandler
 import org.WenuLink.adapters.MissionHandler
+import org.WenuLink.adapters.camera.CameraHandler
 import org.WenuLink.sdk.FCManager
 import org.WenuLink.sdk.MissionActionManager
 import org.WenuLink.sdk.MissionManager
