+++ app/src/main/java/org/WenuLink/adapters/aircraft/TelemetryData.kt
@@ -2,10 +2,10 @@ package org.WenuLink.adapters.aircraft
 
 import com.MAVLink.enums.CAMERA_MODE
 import dji.common.remotecontroller.HardwareState.FlightModeSwitch
-import org.WenuLink.adapters.MessageUtils
 import kotlin.Int
 import kotlin.math.roundToInt
 import kotlin.math.sqrt
+import org.WenuLink.adapters.MessageUtils
 
 /**
  * Data class to hold telemetry info.
@@ -159,49 +159,6 @@ enum class SensorState {
 data class IMUState(
     val gyroscope: MutableList<SensorState> = mutableListOf(),
     val accelerometer: MutableList<SensorState> = mutableListOf()
-)
-
-enum class CameraCaptureType {
-    UNSET,
-    IMAGE,
-    VIDEO
-}
-
-enum class CameraCaptureStatus(val value: Int) {
-    IDLE(0),
-    IN_PROGRESS(1),
-    INTERVAL_IDLE(2),
-    INTERVAL_PROGRESS(3)
-}
-
-data class CameraState(
-    val mavlinkMode: Int = CAMERA_MODE.CAMERA_MODE_IMAGE,
-    val captureType: CameraCaptureType = CameraCaptureType.IMAGE,
-    val captureStatus: CameraCaptureStatus = CameraCaptureStatus.IDLE,
-    /**
-     * captureTime depends on CaptureType:
-     * - CaptureType.IMAGE: capture interval in seconds
-     * - CaptureType.VIDEO: elapsed recording time in milliseconds
-     */
-    val captureTime: Long = 0
-)
-
-data class CameraMetadata(
-    val id: Int = 1,
-    val streamID: String,
-    val name: String,
-    val fwVersion: String,
-    val width: Int,
-    val height: Int,
-    val fps: Int,
-    val state: CameraState = CameraState()
-)
-
-data class ImageMetadata(
-    val index: Int,
-    val captureOk: Boolean,
-    val cameraID: Int,
-    val telemetry: TelemetryData
 )
 
 data class Quaternion(val w: Double, val x: Double, val y: Double, val z: Double) {
