+++ app/src/main/java/org/WenuLink/adapters/camera/CameraHandler.kt
@@ -1,4 +1,4 @@
-package org.WenuLink.adapters
+package org.WenuLink.adapters.camera
 
 import com.MAVLink.enums.CAMERA_MODE
 import io.getstream.log.taggedLogger
@@ -8,11 +8,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.launch
-import org.WenuLink.adapters.actions.CameraCommand
-import org.WenuLink.adapters.aircraft.CameraCaptureStatus
-import org.WenuLink.adapters.aircraft.CameraCaptureType
-import org.WenuLink.adapters.aircraft.CameraMetadata
-import org.WenuLink.adapters.aircraft.CameraState
+import org.WenuLink.adapters.AsyncUtils
 import org.WenuLink.sdk.CameraManager
 
 class CameraHandler {
